// Bi-Cycle-Store/src/app/modules/Orders/order.interface.ts
import  { Types } from "mongoose";
import { Document } from "mongoose";

export interface Order extends Document {
    user: Types.ObjectId;
    products: {
        product: Types.ObjectId;
        quantity: number;

    },
    status: "Pending" | "Paid" 
    orderStatus: 'Pending' | "Accepted"
    totalPrice: number,
    transaction: {
        id: string;
        transactionStatus: string;
        bank_status: string;
        sp_code: string;
        sp_message: string;
        method: string;
        date_time: string;
      };

}